+++ src/definition.ts
@@ -1,6 +1,6 @@
 import * as Data from "effect/Data"
 import type ts from "typescript"
-import * as Nano from "./utils/Nano.js"
+import * as Nano from "./core/Nano.js"
 import type * as TypeCheckerApi from "./utils/TypeCheckerApi.js"
 import type * as TypeScriptApi from "./utils/TypeScriptApi.js"
