+++ src/diagnostics/missingEffectContext.ts
@@ -1,9 +1,9 @@
 import * as ReadonlyArray from "effect/Array"
 import { pipe } from "effect/Function"
 import type ts from "typescript"
+import * as Nano from "../core/Nano.js"
 import type { ApplicableDiagnosticDefinition } from "../definition.js"
 import { createDiagnostic } from "../definition.js"
-import * as Nano from "../utils/Nano.js"
 import * as TypeCheckerApi from "../utils/TypeCheckerApi.js"
 import * as TypeParser from "../utils/TypeParser.js"
 import * as TypeScriptApi from "../utils/TypeScriptApi.js"
