+++ src/index.ts
@@ -3,12 +3,12 @@ import * as ReadonlyArray from "effect/Array"
 import { pipe } from "effect/Function"
 import * as Option from "effect/Option"
 import type ts from "typescript"
+import * as Nano from "./core/Nano.js"
 import { PluginOptions } from "./definition.js"
 import { diagnostics } from "./diagnostics.js"
 import { dedupeJsDocTags, prependEffectTypeArguments } from "./quickinfo.js"
 import { refactors } from "./refactors.js"
 import * as AST from "./utils/AST.js"
-import * as Nano from "./utils/Nano.js"
 import * as TypeCheckerApi from "./utils/TypeCheckerApi.js"
 import * as TypeScriptApi from "./utils/TypeScriptApi.js"
