+++ src/quickinfo.ts
@@ -3,8 +3,8 @@ import * as Eq from "effect/Equivalence"
 import { pipe } from "effect/Function"
 import * as Option from "effect/Option"
 import type ts from "typescript"
+import * as Nano from "./core/Nano.js"
 import * as AST from "./utils/AST.js"
-import * as Nano from "./utils/Nano.js"
 import * as TypeCheckerApi from "./utils/TypeCheckerApi.js"
 import * as TypeParser from "./utils/TypeParser.js"
 import * as TypeScriptApi from "./utils/TypeScriptApi.js"
