+++ src/refactors/removeUnnecessaryEffectGen.ts
@@ -1,8 +1,8 @@
 import { pipe } from "effect/Function"
 import * as Option from "effect/Option"
+import * as Nano from "../core/Nano.js"
 import { createRefactor, RefactorNotApplicableError } from "../definition.js"
 import * as AST from "../utils/AST.js"
-import * as Nano from "../utils/Nano.js"
 import * as TypeParser from "../utils/TypeParser.js"
 import * as TypeScriptApi from "../utils/TypeScriptApi.js"
