+++ src/utils/TypeParser.ts
@@ -1,7 +1,7 @@
 import * as Data from "effect/Data"
 import * as Option from "effect/Option"
 import type ts from "typescript"
-import * as Nano from "./Nano.js"
+import * as Nano from "../core/Nano.js"
 import * as TypeCheckerApi from "./TypeCheckerApi.js"
 import * as TypeScriptApi from "./TypeScriptApi.js"
